requirements.txt
discord.py
python-dotenv
pillow
arabic-reshaper
python-bidi
quart
watchdog
aiohttp
PyNaCl
pytz